sample-project/pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.1.0"
description = "Your project description."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
python-dotenv = "~1.0"

[tool.poetry.group.dev.dependencies]
functions-framework = "~3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"